move package functions to merging logic
