+++ casper/src/main/scala/coop/rchain/casper/merging/ConflictSetMerger.scala
@@ -4,6 +4,7 @@ import cats.effect.Concurrent
 import cats.syntax.all._
 import coop.rchain.rspace.HotStoreTrieAction
 import coop.rchain.rspace.hashing.Blake2b256Hash
+import coop.rchain.rspace.merger.MergingLogic._
 import coop.rchain.rspace.merger.StateChange._
 import coop.rchain.rspace.merger._
 import coop.rchain.shared.{Log, Stopwatch}
